fix: move thunkActions into new redux shape
