+++ src/containers/ReviewActions/components/SubmissionNavigation.jsx
@@ -6,8 +6,7 @@ import { Icon, IconButton } from '@edx/paragon';
 import { ChevronLeft, ChevronRight } from '@edx/paragon/icons';
 import { FormattedMessage, injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 
-import { selectors } from 'data/redux';
-import thunkActions from 'data/thunkActions';
+import { selectors, thunkActions } from 'data/redux';
 import messages from './messages';
 
 /**
